+++ src/PublicAPI.php
@@ -8,20 +8,14 @@ use Klaviyo\Model\ProfileModel;
 class PublicAPI extends KlaviyoAPI
 {
     /**
-     * Track Class constants
-     */
-    const TRACK = 'track';
-    const IDENTIFY = 'identify';
-
-    /**
      * PublicAPI constructor.
      * @param $public_key
      * @param $private_key
      * @param string $host
      */
-    public function __construct( $public_key, $private_key, $host = self::BASE_URL )
+    public function __construct($public_key, $private_key, $host = self::BASE_URL)
     {
-        parent::__construct( $public_key, $private_key, $host = self::BASE_URL );
+        parent::__construct($public_key, $private_key, $host = self::BASE_URL);
     }
 
     /**
@@ -31,11 +25,10 @@ class PublicAPI extends KlaviyoAPI
      * @param EventModel $event
      * @return mixed
      */
-    public function track( EventModel $event )
+    public function track(EventModel $event, $post = false)
     {
-        $options = array( self::QUERY => $event->toArray() );
-
-        return $this->publicRequest( self::TRACK, $options );
+        $options = $post ? $this->createRequestJson($event->toArray()) : $this->createOptionsArray(self::TRACK, $event);
+        return $this->publicRequest(self::TRACK, $options, $post);
     }
 
     /**
@@ -45,14 +38,9 @@ class PublicAPI extends KlaviyoAPI
      * @param ProfileModel $profile
      * @return mixed
      */
-    public function identify( ProfileModel $profile )
+    public function identify(ProfileModel $profile, $post = false)
     {
-        $options = array(
-            self::QUERY => array(
-                self::PROPERTIES => $profile
-            )
-        );
-
-        return $this->publicRequest( self::IDENTIFY, $options );
+        $options = $post ? $this->createRequestJson([self::PROPERTIES => $profile->toArray()]) : $this->createOptionsArray(self::IDENTIFY, $profile);
+        return $this->publicRequest(self::IDENTIFY, $options, $post);
     }
 }